requirements.txt
aiohttp==3.6.2
apturl==0.5.2
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.3
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
Cybernator==0.3
DateTime==4.3
dbus-python==1.2.16
defer==1.0.6
discord.py==1.3.3
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
fastapi==0.58.1
ffmpeg==1.4
flake8==3.8.3
Flask==1.1.2
Flask-API==2.0
httplib2==0.14.0
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
meson==0.53.2
more-itertools==4.2.0
multidict==4.7.6
nest-asyncio==1.3.3
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
Paginator==0.5.1
parso==0.7.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
pydantic==1.5.1
pydocstyle==5.0.2
pyflakes==2.2.0
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.5.2
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.4.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-jsonrpc-server==0.3.4
python-language-server==0.33.3
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
rope==0.17.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
snowballstemmer==2.0.0
starlette==0.13.4
systemd-python==234
testresources==2.0.1
texttable==1.6.2
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
urllib3==1.25.9
wadllib==1.3.3
websocket-client==0.53.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
yapf==0.30.0
yarl==1.4.2
youtube-dl==2020.6.6
zipp==1.0.0
zope.interface==5.1.0
